// src/Types/SessionStatuses.ts
export enum SessionStatuses {
  AGREEMENT = 'AGREEMENT',
  CREATED = 'CREATED',
  ASKING_FOR_PLACE = 'ASKING_FOR_PLACE',
  CHOOSING_PLACE = 'CHOOSING_PLACE',
  ASKING_FOR_COMMENT = 'ASKING_FOR_COMMENT',
  REQUESTING_SERVICE = 'REQUESTING_SERVICE',
  SERVICE_IN_PROGRESS = 'SERVICE_IN_PROGRESS',
  COMPLETED = 'COMPLETED',
  ASKING_FOR_NAME = 'ASKING_FOR_NAME',

}